# swagger.yaml
swagger: "2.0"
info:
  description: "An API which facilitates the creation, modification, and deletion of bundles.  A bundle is a collection of datasets which are to be published at the same date and time."
  version: "1.0.0"
  title: "Bundle Administration"
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
basePath: "/v1"
...
schemes:
  - "http"
security:
  - Authorization: []
parameters:
...
  bundle:
    required: true
    name: bundle
    schema:
      $ref: "#/definitions/Bundle"
    description: "The bundle definition"
    in: body
...
  limit:
    name: limit
    description: "Maximum number of items that will be returned. A value of zero will return zero items."
    in: query
    required: false
    type: integer
    default: 20
    minimum: 0
    maximum: 1000
  offset:
    name: offset
    description: "Starting index of the items array that will be returned. By default it is zero, meaning that the returned items will start from the beginning."
    in: query
    required: false
    type: integer
    default: 0
    minimum: 0
  publish_date:
    name: publish_date
    description: "Filter bundles by their scheduled publication date. Accepts an optional datetime value and returns all bundles where the scheduled_at field matches the specified datetime."
    in: query
    required: false
    type: string
    format: date-time
...
securityDefinitions:
  Authorization:
    name: Authorization
    description: "Access token provided by Auth Service in a Bearer format. Can be a human or service user token."
    in: header
    type: apiKey
paths:
  /bundles:
    get:
      tags:
        - "Private"
      summary: "List bundles"
      description: "Returns a list of all bundles available."
      parameters:
        - $ref: "#/parameters/limit"
        - $ref: "#/parameters/offset"
        - $ref: "#/parameters/publish_date"
      produces:
        - "application/json"
      responses:
        200:
          description: "A json list containing bundles"
          headers:
            ETag:
              description: The RFC9110 ETag header field. Defines the unique entity tag for the current state of the resource. This is used for setting the `If-Match` and `If-None-Match` headers on subsequent requests.
              type: string
              pattern: ^(?:W/)?"(?:[!#-~])+"$
            Cache-Control:
              description: The RFC9111 Cache-Control header field for the response which instructs how to handle caching the resource.
              type: string
          schema:
            $ref: "#/definitions/Bundles"
        400:
          $ref: "#/responses/InvalidRequest"
        401:
          $ref: "#/responses/UnauthorisedError"
        403:
          $ref: "#/responses/ForbiddenError"
        500:
          $ref: "#/responses/InternalError"
    post:
      tags:
        - "Private"
      summary: "Create a bundle"
      description: "Creates a bundle in the database which groups datasets together to be published on the same date and time"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - $ref: "#/parameters/bundle"
      responses:
        201:
          description: "bundle was created"
          headers:
            ETag:
              description: The RFC9110 ETag header field. Defines the unique entity tag for the current state of the resource. This is used for setting the `If-Match` and `If-None-Match` headers on subsequent requests.
              type: string
              pattern: ^(?:W/)?"(?:[!#-~])+"$
            Cache-Control:
              description: The RFC9111 Cache-Control header field for the response which instructs how to handle caching the resource.
              type: string
            Location:
              description: The RFC9110 Location header field. Defines the access location (i.e. path) of the primary resource created for use in subsequent requests.
              type: string
          schema:
            $ref: "#/definitions/Bundle"
        400:
          $ref: "#/responses/InvalidRequest"
        401:
          $ref: "#/responses/UnauthorisedError"
        403:
          $ref: "#/responses/ForbiddenError"
        500:
          $ref: "#/responses/InternalError"
...
responses:
...
  ForbiddenError:
    description: Access denied.
    schema:
      $ref: "#/definitions/ErrorList"
  InternalError:
    description: "Failed to process the request due to an internal error"
    schema:
      $ref: "#/definitions/ErrorList"
  InvalidRequest:
    description: Unable to process request due to a malformed or invalid request body or query parameter.
    schema:
      $ref: "#/definitions/ErrorList"
...
  UnauthorisedError:
    description: Authentication information is missing or invalid
    schema:
      $ref: "#/definitions/ErrorList"
definitions:
  Bundles:
    description: "A list of bundles"
    type: object
    allOf:
      - $ref: "#/definitions/PaginationFields"
      - type: object
        properties:
          items:
            type: array
            items:
              $ref: "#/definitions/Bundle"
  Bundle:
    description: "A model for the response body when retrieving a bundle"
    type: object
    required:
      - bundle_type
      - preview_teams
      - title
      - managed_by
      - e_tag
    properties:
      bundle_type:
        description: "The type of bundle."
        type: string
        enum:
          - MANUAL
          - SCHEDULED
        example: SCHEDULED
      created_by:
        description: "The user that created the bundle, taken from the auth token supplied.  This should only be set on initial creation of the bundle."
        readOnly: true
        type: object
        required:
          - email
        properties:
          email:
            description: The email of the user who created the bundle.
            type: string
            example: "publisher@ons.gov.uk"
      created_at:
        description: "The ISO8601 date-time the bundle was created at."
        type: string
        format: date-time
        readOnly: true
        example: "2025-04-04T07:00:00.000Z"
      id:
        type: string
        description: "The bundle id"
        readOnly: true
        minLength: 1
        maxLength: 100
        pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
        example: "9e4e3628-fc85-48cd-80ad-e005d9d283ff"
      last_updated_by:
        description: "The user that last updated the bundle, taken from the auth token supplied."
        readOnly: true
        type: object
        required:
          - email
        properties:
          email:
            description: The email of the user who updated the bundle.
            type: string
            example: "publisher@ons.gov.uk"
      preview_teams:
        description: "A list of teams who have permissions to view the dataset series in the bundle."
        type: array
        minLength: 0
        items:
          type: object
          required:
            - id
          properties:
            id:
              description: The preview team ID.
              type: string
              minLength: 1
              pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
              example: 1253e849-01fd-4662-bee2-63253538da93
      scheduled_at:
        description: "The ISO8601 date-time the bundle is scheduled to publish at."
        type: string
        format: date-time
        example: "2025-04-04T07:00:00.000Z"
      state:
        $ref: "#/definitions/BundleState"
      title:
        type: string
        description: "The title of the bundle"
        minLength: 1
        example: "CPI March 2025"
      updated_at:
        description: "The ISO8601 date-time the bundle was last updated at."
        type: string
        readOnly: true
        format: date-time
        example: "2025-04-04T07:00:00.000Z"
      managed_by:
        description: "The system that created and manages the bundle"
        type: string
        enum:
          - WAGTAIL
          - DATA-ADMIN
        example: WAGTAIL
      e_tag:
        description: "Auto generated ETag that defines the unique entity tag for the current state of the resource."
        type: string
        example: "c7e4b9a2f813d6e5f0a9d3c1e7f8b4a5d6c7e9f0"
...
  ErrorList:
    description: "A list of errors that occurred."
    type: object
    properties:
      errors:
        type: array
        items:
          $ref: "#/definitions/Error"
  Error:
    description: "The details of a specific error."
    type: object
    properties:
      code:
        type: string
        description: "Code representing the type of error that occurred for use by consuming systems to identify error types."
        enum:
          - SomeError # TODO: Replace with list of actual error codes once defined
          - AnotherError
      description:
        type: string
        description: "Human readable description of the error"
      source:
        description: The details of which field or parameter the error relates to. Used to return validation errors to 4xx requests. Only one of the properties below can be returned in any single error.
        type: object
        properties:
          field:
            description: The RFC 6901 JSON Pointer to the JSON field in the request body to which the error applies.
            type: string
            example: "/preview_teams/0"
          parameter:
            description: The path or query parameter to which the error applies.
            type: string
            example: "limit"
          header:
            description: The header to which the error applies.
            type: string
            example: If-Match
...
  PaginationFields:
    type: object
    properties:
      count:
        description: "The number of items returned."
        readOnly: true
        type: integer
        example: 20
      limit:
        description: "The number of items requested."
        type: integer
        default: 20
        minimum: 0
        maximum: 1000
      offset:
        description: "The offset of the first item to retrieve, starting at 0. Use this parameter as a pagination mechanism along with the limit parameter"
        type: integer
        example: 0
        default: 0
      total_count:
        description: "The total number of items available."
        readOnly: true
        type: integer
        example: 123
  BundleState:
    description: |
      The current workflow state of the bundle as a whole.

      The possible states are:
        * `DRAFT `: Bundle has been created and is being populated with content changes.
        * `IN_REVIEW`: Bundle has been submitted for review.
        * `APPROVED`: Bundle has been approved and is awaiting release.
        * `PUBLISHED `: Bundle has been published and is now public.
    type: string
    enum:
      - DRAFT
      - IN_REVIEW
      - APPROVED
      - PUBLISHED
    example: APPROVED
    default: DRAFT